% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_check.R
\name{herb_check}
\alias{herb_check}
\title{Checking the main parameters
\code{link{sel_herb}}}
\usage{
herb_check(put, sdrate, thresh, rate)
}
\arguments{
\item{put}{probability of a weed to be untouched by the herbicide. \code{numeric}, \eqn{0 \le \code{put} \le 1}.}

\item{sdrate}{variance of the herbicide rate reaching the weed. \code{positive numeric}, 1 = 1 unit standard deviation.}

\item{thresh}{threshold herbicide rate to kill weeds without resistance. \code{numeric}, \eqn{0 \le \code{thresh} \le 1}.}

\item{rate}{percentage (\%) of the registered herbicide dose. \code{positive numeric}, can exceed 100 \%.}
}
\description{
Is called within \code{sel_herb()} to check for the existence of the necessary data.
}
