% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Klasse-Methode-PROSPER.R
\name{intern_SimData<-}
\alias{intern_SimData<-}
\title{replacement of Simulation Data}
\usage{
intern_SimData(object) <- value
}
\arguments{
\item{object}{\code{prosper}, the simulation data of this object replaced.}

\item{value}{object fitting the requirements for simulation data of the \code{class prosper}, this replaces the existing simulation data.}
}
\value{
object of class \code{prosper}
}
\description{
\code{intern_SimData<-} replaces the simulation data of a \code{prosper}-object.
}
\keyword{internal}
