% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Klasse-Methode-PROSPER.R
\name{intern_SimData}
\alias{intern_SimData}
\alias{intern_SimData,prosper-method}
\title{Extraction Simulation Data}
\usage{
intern_SimData(object)

\S4method{intern_SimData}{prosper}(object)
}
\arguments{
\item{object}{\code{prosper}, the simulation data of this object will be extracted.}
}
\value{
a \code{data.table} object
}
\description{
\code{intern_SimData} extracts the simulation data from a \code{prosper}-object
}
\examples{
library(data.table)
struc_preparation2(Rmx=10, af=c(0.01,0.8), epis=0, dom=c(1,0.3))
   #simdata[,"repetition":=1]
   #simdata[,"simcycle":=1]
   #setcolorder(simdata, c(c("repetition", "simcycle"),
   #colnames(simdata)[!(colnames(simdata) \%in\%
   #c("simulation_run", "repetition", "simcycle"))])) 
#prosp_ob <- prosper(simstruc = c(5,10), loci=2, simData = simdata)
#intern_SimData(prosp_ob)
}
\keyword{internal}
