% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Klasse-Methode-PROSPER.R
\name{struc_addSimData}
\alias{struc_addSimData}
\alias{struc_addSimData,prosper-method}
\title{Wrapper for \code{intern_SimData<-}  generic}
\usage{
struc_addSimData(object, value, add = TRUE)

\S4method{struc_addSimData}{prosper}(object, value, add = TRUE)
}
\arguments{
\item{object}{a \code{prosper} object}

\item{value}{code{data.frame} or code{data.table}}

\item{add}{\code{logical}, when \code{TRUE} adding data to the existing data. Otherwise the value is completely new}
}
\description{
This function ensures the correct form of data, before adding data of a new simulation cycle to a \code{prosper} object using \code{intern_SimData}.
}
\keyword{internal}
