% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc_endSim.R
\name{struc_endSim}
\alias{struc_endSim}
\title{Check for some stop conditions.}
\usage{
struc_endSim(simcycle = year, break_col_names = "SB_autumn")
}
\arguments{
\item{simcycle}{The count of simulation runs. The  value is \code{duration} \eqn{\ge  0}.}

\item{break_col_names}{The names of the columns in \code{dfgenotype}, which are the first in the next simulation cycle.}
}
\value{
\code{logical} TRUE if the simulation has to stop.
}
\description{
Conditions that are necessary to continue the simulation are checked.
}
\details{
This function combines two conditions to terminate the simulation. The first is the \code{duration} that is defined for the simulation. The second is the the number of individuals for the next simulation cycle: when the population is extinct the simulation ends.
}
\seealso{
\code{\link{struc_saveSimData}}
}
