% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struc_nextCycle.R
\name{struc_nextCycle}
\alias{struc_nextCycle}
\title{Preparing the data for the next simulation run}
\usage{
struc_nextCycle(start_name, end_name)
}
\arguments{
\item{start_name}{name of the column with the start population for the next simulation cycle. If missing the column name after the column 'resist' will be used. \code{character}.}

\item{end_name}{name of the column with the population at the end of the simulation cycle. This will be used as the start values in the next simulation cycle (in the column with \code{start_name}). If missing the last column name is used. \code{character}.}
}
\value{
A \code{data.table} object.
}
\description{
\code{struc_nextCycle} finishes the simulation cycle. Therefore, the \code{dfgenotype} is reset and prepared for the next cycle.
}
\seealso{
\code{\link{struc_endSim}} \code{\link{struc_saveSimData}}
}
